+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [11.3.0] - 2025-04-23
+
 ### Fixed
 
 - Can't select text when `readOnly` is true [#2529](https://github.com/singerdmx/flutter-quill/pull/2529).
@@ -142,7 +144,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Apple-specific font dependency for subscript and superscript functionality from the example.
 - **BREAKING**: The [`super_clipboard`](https://pub.dev/packages/super_clipboard) plugin, To restore legacy behavior for `super_clipboard`, use [`flutter_quill_extensions`](https://pub.dev/packages/flutter_quill_extensions) package and `FlutterQuillExtensions.useSuperClipboardPlugin()`.
 
-[unreleased]: https://github.com/singerdmx/flutter-quill/compare/v11.2.0...HEAD
+[unreleased]: https://github.com/singerdmx/flutter-quill/compare/v11.3.0...HEAD
+[11.3.0]: https://github.com/singerdmx/flutter-quill/compare/v10.0.0...v11.3.0
 [11.2.0]: https://github.com/singerdmx/flutter-quill/compare/v10.0.0...v11.2.0
 [11.1.2]: https://github.com/singerdmx/flutter-quill/compare/v10.0.0...v11.1.2
 [11.1.1]: https://github.com/singerdmx/flutter-quill/compare/v10.0.0...v11.1.1